moved tic and toc from Quality to TestCase and bugfix This prepares the next step of development of Benchmarker, adding a stack. To this end, toc must return all vital pieces of information. But this information must be accessible to TestCase, so invocation is done best in TestCase also.
Moving all benchmarking from Quality to TestCase, 
we observed a bug: Quality.setFinished() is invoked in any case a testcase finishes, whether successfully, 
with error or failure and also with invalid assertion. 
Thus the `assert false` had to be removed again. 
For benchmarking this means, that only Quality.setFinished shall have a toc. 
Moving the toc to TestCase, accordingly. 

In addition, several minor improvements are done.